+++ Attrib/Attrib.cpp
@@ -1,5 +1,4 @@
 
-#include "stdafx.h"
 #include "Attrib.h"
 
 namespace Attrib
