+++ NFSMW_XenonEffects.vcxproj
@@ -53,7 +53,7 @@
   </PropertyGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">
     <ClCompile>
-      <PrecompiledHeader>Use</PrecompiledHeader>
+      <PrecompiledHeader>NotUsing</PrecompiledHeader>
       <WarningLevel>Level3</WarningLevel>
       <Optimization>Disabled</Optimization>
       <PreprocessorDefinitions>WIN32;_DEBUG;_WINDOWS;_USRDLL;_CRT_SECURE_NO_WARNINGS;%(PreprocessorDefinitions)</PreprocessorDefinitions>
@@ -72,7 +72,7 @@
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">
     <ClCompile>
       <WarningLevel>Level4</WarningLevel>
-      <PrecompiledHeader>Use</PrecompiledHeader>
+      <PrecompiledHeader>NotUsing</PrecompiledHeader>
       <Optimization>MaxSpeed</Optimization>
       <FunctionLevelLinking>true</FunctionLevelLinking>
       <IntrinsicFunctions>true</IntrinsicFunctions>
@@ -93,7 +93,13 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClInclude Include="Attrib\Attrib.h" />
-    <ClInclude Include="stdafx.h" />
+    <ClInclude Include="Attrib\AttribArray.h" />
+    <ClInclude Include="Attrib\AttribAttribute.h" />
+    <ClInclude Include="Attrib\AttribCollection.h" />
+    <ClInclude Include="Attrib\AttribCore.h" />
+    <ClInclude Include="Attrib\AttribHashMap.h" />
+    <ClInclude Include="Attrib\AttribInstance.h" />
+    <ClInclude Include="Attrib\AttribNode.h" />
     <ClInclude Include="targetver.h" />
     <ClInclude Include="UMath\UMath.h" />
   </ItemGroup>
@@ -106,10 +112,6 @@
       <CompileAsManaged Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">false</CompileAsManaged>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">
       </PrecompiledHeader>
-    </ClCompile>
-    <ClCompile Include="stdafx.cpp">
-      <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">Create</PrecompiledHeader>
-      <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">Create</PrecompiledHeader>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
